requirements.txt
Django==6.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
django-crispy-forms==2.1
